build system: add several more warning flags & fix results

+++ src/common/math.cpp
@@ -44,7 +44,7 @@ int_log2(uint64_t value) {
 double
 int_to_double(int64_t value) {
   if (static_cast<uint64_t>(value + value) > (0xffeull << 52))
-    return NAN;
+    return static_cast<double>(NAN);
   return std::ldexp(((value & ((1ll << 52) - 1)) + (1ll << 52)) * (value >> 63 | 1), (value >> 52 & 0x7ff) - 1075);
 }
